Authorizer: Migration script (#1173)
* authorizer: prepare migration script

* authorizer: fix fork test

* authorizer: fix fork test

* authorizer: update migration with latest model changes

* authorizer: delay root change

* authorizer: test partial migrations

* authorizer: fix linter

* authorizer: improve deployment input layout
