+++ pkg/deployments/tasks/2022xxxx-timelock-authorizer/readme.md
@@ -1,0 +1,9 @@
+# 2022-xx-xx - Timelock Authorizer
+
+Deployment of the `TimelockAuthorizer` in order to replace the current authorizer of the Vault.
+
+## Useful Files
+
+- [Ethereum mainnet addresses](./output/mainnet.json)
+- [`TimelockAuthorizer` ABI](./abi/TimelockAuthorizer.json)
+- [`TimelockAuthorizerMigrator` ABI](./abi/TimelockAuthorizerMigrator.json)
