+++ pkg/vault/contracts/authorizer/IBasicAuthorizer.sol
@@ -1,0 +1,26 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+pragma solidity ^0.7.0;
+
+import "../interfaces/IAuthorizer.sol";
+
+interface IBasicAuthorizer is IAuthorizer {
+    // solhint-disable-next-line func-name-mixedcase
+    function DEFAULT_ADMIN_ROLE() external returns (bytes32);
+
+    function getRoleAdmin(bytes32 role) external view returns (bytes32);
+
+    function getRoleMember(bytes32 role, uint256 index) external view returns (address);
+
+    function getRoleMemberCount(bytes32 role) external view returns (uint256);
+}
