+++ pkg/vault/contracts/test/MockBasicAuthorizer.sol
@@ -1,0 +1,79 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+pragma solidity ^0.7.0;
+pragma experimental ABIEncoderV2;
+
+import "@balancer-labs/v2-solidity-utils/contracts/helpers/BalancerErrors.sol";
+import "@balancer-labs/v2-solidity-utils/contracts/helpers/InputHelpers.sol";
+import "@balancer-labs/v2-solidity-utils/contracts/openzeppelin/EnumerableSet.sol";
+
+import "../authorizer/IBasicAuthorizer.sol";
+
+contract MockBasicAuthorizer is IBasicAuthorizer {
+    using EnumerableSet for EnumerableSet.AddressSet;
+
+    bytes32 public constant override DEFAULT_ADMIN_ROLE = 0x00;
+
+    struct RoleData {
+        EnumerableSet.AddressSet members;
+        bytes32 adminRole;
+    }
+
+    mapping(bytes32 => RoleData) private _roles;
+
+    constructor() {
+        _grantRole(DEFAULT_ADMIN_ROLE, msg.sender);
+    }
+
+    function hasRole(bytes32 role, address account) public view returns (bool) {
+        return _roles[role].members.contains(account);
+    }
+
+    function canPerform(
+        bytes32 actionId,
+        address account,
+        address
+    ) external view override returns (bool) {
+        return hasRole(actionId, account);
+    }
+
+    function getRoleMemberCount(bytes32 role) external view override returns (uint256) {
+        return _roles[role].members.length();
+    }
+
+    function getRoleMember(bytes32 role, uint256 index) external view override returns (address) {
+        return _roles[role].members.at(index);
+    }
+
+    function getRoleAdmin(bytes32 role) external view override returns (bytes32) {
+        return _roles[role].adminRole;
+    }
+
+    function grantRolesToMany(bytes32[] memory roles, address[] memory accounts) external {
+        InputHelpers.ensureInputLengthMatch(roles.length, accounts.length);
+        for (uint256 i = 0; i < roles.length; i++) {
+            grantRole(roles[i], accounts[i]);
+        }
+    }
+
+    function grantRole(bytes32 role, address account) public {
+        _require(hasRole(_roles[role].adminRole, msg.sender), Errors.GRANT_SENDER_NOT_ADMIN);
+        _grantRole(role, account);
+    }
+
+    function _grantRole(bytes32 role, address account) private {
+        _roles[role].members.add(account);
+    }
+}
